Introduce yang-lsp-settings-schema.json
The JSON Schema for the yang.settings file is specified to allow for
various IDE plugins to take the LSP version corresponding schema and
associate them to the yang.settings for ease of editing and review.

It is also updated that yang.settings is a JSON with Comments filetype
since it was already supporting comments.

This change addresses the issue TypeFox/yang-lsp#233

Signed-off-by: Siddharth Sharma <siddharth.sharma@ericsson.com>